requirements.txt
Pillow
matplotlib
numpy
click
coverage
docker
flask
furl
requests
Sphinx
termcolor
pika
protobuf
psycopg2-binary
requests-mock
sqlalchemy